[Circuit] Separate out core classes from elements
